// frontend/src/data/models/error-codes.ts
export enum ErrorCodes {
  NON_FILLED_FIELD = "Необходимо заполнить поле",
  EMAIL_BIGGER = "Длина почтового адреса не может быть больше 256 символов",
  INCORRECT_EMAIL = "Некорректный формат адреса почты",
  TAKEN_EMAIL = "Такая почта уже занята",
  TAKEN_NICKNAME = "Такой никнейм уже занят",
  NICKNAME_BIGGER = "Длина никнейма не может быть больше 20 символов",
  NICKNAME_LESS = "Длина никнейма не может быть меньше 4 символов",
  NICKNAME_LOGIN = "Некорректные символы в никнейме. Допустимые символы: прописные, строчные буквы латиницы, цифры, знаки (., _, -)",
  PASSWORDS_MISMATCH = "Введенные пароли не совпадают",
  PASSWORD_LESS = "Длина пароля не может быть меньше 8 символов",
  PASSWORD_BIGGER = "Длина пароля не может быть больше 72 символов",
  LOGIN_ERROR = "Пользователь с таким логином или паролем не найден",
  INCORRECT_NAME = "Некорректные символы в имени. Допустимые символы: прописные, строчные буквы латиницы",
  NAME_BIGGER = "Длина имени не может быть больше 70 символов",
  INCORRECT_SURNAME = "Некорректные символы в фамилии. Допустимые символы: прописные, строчные буквы латиницы",
  SURNAME_BIGGER = "Длина фамилии не может быть больше 70 символов",
  INCORRECT_OLD_PASSWORD = "Такого пароля нет"
}